Renaming PerformanceView to GUI
